Limit concurrency of jobs by cancelling or queueing as appropriate (#16)
This now will cancel any ongoing 'style_and_document' job when another one starts. This should reduce the number of unnecessary jobs when multiple commits are made in quick succession.

There is also now an option `limit_parallel` which isn't required and is `false` by default, if it's set to `true` it will force any R-CMD-CHECK or test-coverage jobs to queue up. The aim here is to limit the number of tests running simultaneously which is useful when they call an API etc, that might not handle multiple simultaneous requests well.

To use the new argument `phs_checks.yaml` should be updated as below:

```yaml
jobs:
  PHS-checks:
    uses: Public-Health-Scotland/actions/.github/workflows/phs_package_checks.yaml@v1.5.0
    with:
      limit_parallel: true
    permissions: 
      contents: write
      pull-requests: write
    secrets: inherit
```