Add pkgdown site build and deploy to the package checks flow (#22)
This adds a job to the workflow to build and deploy a [pkgdown](https://pkgdown.r-lib.org/) site. If the package is already set up for this, the existing pkgdown workflow can be removed. If the packager doesn't have pkgdown yet, this should still work but won't deploy a site. You must create a `_pkgdown.yaml` file, best using `usethis::use_pkgdown()`, and configure the repository settings to deploy a site from the `gh-pages` branch.