Fix commit permission issues. (#73)
Resolve the issue with the CI failing if it doesn't have permission to commit to a branch - This is the case if the branch is protected, or if the PR is coming from a branch on a fork.

Now the CI will always continue (so run the R-CMD-checks etc.) and if it can't commit, it will open a PR with the suggested changes instead.